Added files via upload.
Site is moving along as planned. Various fixes and updates have been outlined and noted to client, such as:

styling the homepage text content 
links changing colour on hover
full mobile "responsiveness"
small mobile interactive elements as necessary
updating fonts to a stronger/more cohesive serif choice
updating font sizes for readability across browser
photos!
fixing the navigation bar at the top (desktop) and hiding it in the mobile view
increasing the opacity of the navigation bar on scrolling (as to not confuse site readers!)
implementing the mobile navigation bar and hiding it in the desktop view
adding the various meta descriptions and content that search engines will use to index your site and provide details for search engine optimization
uploading/adding your articles for viewing on the "Colleagues" page

All of these are forthcoming and should be fixed in subsequent uploads/fixes.