Add files via upload
Completely overwrote code with content updates, as well as modified styles. Fixes of the following were made, too:

links changing colour on hover
updated font sizes for readability
adding more "room" between certain text content for more readability
fixing the position of the navigation bar at the top (in desktop)
as well, giving it a slightly darker background and different opacity and border to help it stand out for site visitors while scrolling (in desktop)
adding meta descriptions for site users for social websites like Twitter
increased font sizes for readability
changed the font choice for the navigation bar (desktop) for the user experience

As well, the site is now linked to a domain.